[CI experiment] revert MOVA audio_vae_precision bf16 -> fp32
Single-variable test to confirm PR #25411 (commit 596b45b37) is the
trigger for ci-monitor cluster R53 (HIPRTC uint32_t on DAC snake in
multimodal-gen-test-2-gpu-amd shard 1). Reverts only the
audio_vae_precision change; leaves vae_precision (video Wan VAE) as
bf16 per PR #25411. DO NOT MERGE.

+++ python/sglang/multimodal_gen/configs/pipeline_configs/mova.py
@@ -40,7 +40,11 @@ class MOVAPipelineConfig(PipelineConfig):
     vae_config: WanVAEConfig = field(default_factory=WanVAEConfig)
     vae_precision: str = "bf16"
     audio_vae_config: DacVAEConfig = field(default_factory=DacVAEConfig)
-    audio_vae_precision: str = "bf16"
+    # TEMP REVERT for CI causation experiment: PR #25411 changed this from
+    # "fp32" -> "bf16" on 2026-05-16; cluster R53 (HIPRTC uint32_t on DAC
+    # snake) started failing immediately after. This branch flips just this
+    # one line back to "fp32" to confirm causation. DO NOT MERGE.
+    audio_vae_precision: str = "fp32"
 
     # Text encoder (UMT5 compatible)
     text_encoder_configs: tuple = field(default_factory=lambda: (T5Config(),))
